requirements.txt
streamlit==1.32.0
google-generativeai==0.3.1
google-api-python-client==2.108.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.0.1
plotly==5.18.0
altair==5.2.0
networkx==3.2.1
python-dotenv==1.0.0